Add an option to refresh the lock automatically
This introduces a new parameter `refresh_interval` which, when set to a
value > 0 triggers the creation of a subthread that will re-set the
expire time of the lock.

This allows one to use a relatively low expires time but keep the lock
active in Redis for as long as the lock is acquired in Python code,
avoiding the stale lock issue when a Python process is killed abruptly
without a chance to release the lock in Redis.